refactor: Clean up project structure and improve speech recognition routing
- Fix speech-to-text routing to populate form fields only, not bottom input
- Add word-to-number conversion for spoken quantities (e.g., 'one hundred' -> '100')
- Add debug logging for speech recognition troubleshooting
- Remove 17 debug/test scripts from backend/scripts/
- Remove 5 redundant documentation files from root and frontend
- Clean up __pycache__ directories
- Remove obsolete chat.html and SETUP_TRIGGER.md
- Improve code organization and maintainability
- Reduce project size by 1730 deleted lines

Speech Recognition Improvements:
- selectedFieldRef correctly set before recognition.start()
- onresult handler routes speech ONLY to target form state
- Word-to-number conversion handles common spoken numbers
- Console logging for debugging speech routing issues

Project Cleanup:
- Deleted deprecated test and debug scripts
- Removed redundant documentation files
- Cleaned build cache directories
- Organized file structure for better maintainability

+++ backend/app/models.py
@@ -118,10 +118,10 @@ class InventoryMaster(Base):
     InventoryID = Column(Integer, primary_key=True, autoincrement=True)
     PID = Column(String(50), ForeignKey("Product_Master.PID"))
     Category = Column(String(100))  # "Produced" or "Sold"
-    QuantityMT = Column(Integer)  # Quantity in Metric Tons (MT)
+    QuantityMT = Column(Float)  # Quantity in Metric Tons (MT) - stored as DECIMAL(18,2) in DB
     ProducedDate = Column(Date)
-    InitialPrice = Column(Integer)  # Price per MT (in currency)
-    SellingPrice = Column(Integer)  # Price per MT (in currency)
+    InitialPrice = Column(Float)  # Price per MT (in currency) - stored as DECIMAL(18,2) in DB
+    SellingPrice = Column(Float)  # Price per MT (in currency) - stored as DECIMAL(18,2) in DB
 
 
 class ComplaintMaster(Base):
